requirements.txt
# Core Dependencies
dnspython>=2.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
jinja2>=3.1.2
colorama>=0.4.6
python-nmap>=0.7.1
validators>=0.22.0
tldextract>=5.1.0
pyyaml>=6.0.1
tabulate>=0.9.0

# Security & Networking
shodan>=1.31.0
python-whois>=0.8.0
builtwith>=1.3.4

# Reporting
markdown>=3.5.0
weasyprint>=60.0

# CLI
click>=8.1.7
rich>=13.7.0

# Utilities
python-dateutil>=2.8.2
ipaddress>=1.0.23
